Pull text-encoding from npm instead of from vtt.js bower dist.
The text-encoding lib (formerly stringencoding) was previously not
on npm and so I elected to pull it from vtt.js bower distributable
instead of using a git submodule, since the current version of it
was broken. It's on npm now so we can pull directly from there and
get rid of our bower dependencies.